+++ lib/src/evaluator.dart
@@ -19,14 +19,14 @@ abstract class PhenotypeEvaluator<P extends Phenotype<G, R>, G,
   /// nothing.
   void destroy() {}
 
-  /// Evaluate takes the phenotype and returns its fitness score. The lower the
+  /// Takes the phenotype and returns its fitness score. The lower the
   /// fitness score, the better the phenotype. Fitness score of [:0.0:] means
   /// that the phenotype is perfect.
   Future<R> evaluate(P phenotype);
 
   /// When evaluators need some work for initialization. By default,
   /// this just returns an immediate [Future].
-  Future init() => Future<Null>.value();
+  Future<void> init() => Future<void>.value();
 }
 
 /// For use when multiple experiments should be done with each phenotype.
@@ -34,7 +34,7 @@ abstract class PhenotypeSerialEvaluator<P extends Phenotype<G, R>, G,
     R extends FitnessResult> extends PhenotypeEvaluator<P, G, R> {
   final FitnessResultCombinator<R> _resultCombinator;
 
-  late R cummulativeResult;
+  late R cumulativeResult;
 
   final R _initialResult;
 
@@ -43,7 +43,7 @@ abstract class PhenotypeSerialEvaluator<P extends Phenotype<G, R>, G,
   @override
   Future<R> evaluate(P phenotype) {
     printf('Evaluating $phenotype');
-    cummulativeResult = _initialResult;
+    cumulativeResult = _initialResult;
     userData = null;
     _completer = Completer<R>();
     _next(phenotype, 0);
@@ -58,16 +58,16 @@ abstract class PhenotypeSerialEvaluator<P extends Phenotype<G, R>, G,
   void _next(P phenotype, int experimentIndex) {
     runOneEvaluation(phenotype, experimentIndex).then((R? result) {
       if (result == null) {
-        printf('Cummulative result for phenotype: $cummulativeResult');
-        _completer.complete(cummulativeResult);
+        printf('Cummulative result for phenotype: $cumulativeResult');
+        _completer.complete(cumulativeResult);
       } else if (result.evaluate()!.isInfinite) {
         printf(
             'Result for experiment #$experimentIndex: FAIL\nFailing phenotype');
-        _completer.complete(cummulativeResult);
+        _completer.complete(cumulativeResult);
       } else {
-        cummulativeResult = _resultCombinator(cummulativeResult, result);
+        cumulativeResult = _resultCombinator(cumulativeResult, result);
         printf(
-            'Result for experiment: $result (cummulative: $cummulativeResult)');
+            'Result for experiment: $result (cummulative: $cumulativeResult)');
         _next(phenotype, experimentIndex + 1);
       }
     });