+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## 0.5.0
- - Null Safety migration
+
+- BREAKING: Null Safety migration
+- BREAKING: fixed typo in `GenerationBreeder.propability` to `.probability`
+- BREAKING: fixed typo in `Evaluator.cummulative` to `.cumulative`
+- BREAKING: fixed typo in `Generation.cummulativeFitnes` to `.cumulativeFitness`
+- BREAKING: changed `ALL_CAPS_FIELDS` into the more idiomatic `camelCaseFields`
+  - `GeneticAlgorithm.MAX_EXPERIMENTS` to `thresholdResult`
+  - `GeneticAlgorithm.MAX_GENERATIONS_IN_MEMORY` to `maxGenerationsInMemory`
 
 ## 0.4.2